backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.1
PyMySQL==1.1.1
python-dotenv==1.0.1
scikit-learn==1.5.0
joblib==1.4.2
numpy==1.26.4
requests==2.32.3
gunicorn==22.0.0